+++ CHANGELOG.md
@@ -4,6 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+### Added
+
+- Mosaic and ImageStack instances now provide a means to change the resampling method.
+
 ## v1.3.0 - 09/24/2024
 
 ### Added